nil getting and setting for all string items works, tests passing
